+++ src/theme/index.ts
@@ -41,6 +41,7 @@ import CustomSlider from "./components/slider";
 import StatusBadge from "./components/statusBadge";
 import StructuredContent from "./components/structuredContent";
 import Tabs from "./components/tabs";
+import CustomTable from "./components/customTable";
 import TemplateStyles from "./components/template";
 import Text from "./components/text";
 import TextInput from "./components/textInput";
@@ -113,6 +114,7 @@ const theme = extendTheme({
     StatusBadge,
     StructuredContent,
     Tabs,
+    CustomTable,
     ...TemplateStyles,
     Text,
     TextInput,