+++ CHANGELOG.md
@@ -8,6 +8,8 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 
 ## Prerelease
 
+## 0.25.9 (February 3, 2022)
+
 ### Adds
 
 - Adds the `Logo` component.